Refactor: move DirectiveGuard to RateLimiting namespace
Update DirectiveGuard namespace to Nalix.Runtime.Internal.RateLimiting and update all references in middleware files. Replace Nalix.Runtime.Pipeline project references with Nalix.Runtime in test projects and solution, removing Pipeline from the solution and build configs.
