+++ tests/Nalix.Tests.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
 VisualStudioVersion = 18.1.11312.151
@@ -20,8 +20,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Framework", "..\src\Nalix.Framework\Nalix.Framework.csproj", "{E7F2E391-534F-4BFC-85D2-0949E87CEE3C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Network", "..\src\Nalix.Network\Nalix.Network.csproj", "{9A2197A7-5108-445B-8B54-4B46E525A7D3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Runtime.Pipeline", "..\src\Nalix.Runtime.Pipeline\Nalix.Runtime.Pipeline.csproj", "{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Runtime.Tests", "Nalix.Runtime.Tests\Nalix.Runtime.Tests.csproj", "{D55E920A-8B56-43AD-A799-6E6154193427}"
 EndProject
@@ -153,18 +151,6 @@ Global
 		{9A2197A7-5108-445B-8B54-4B46E525A7D3}.Release|x64.Build.0 = Release|Any CPU
 		{9A2197A7-5108-445B-8B54-4B46E525A7D3}.Release|x86.ActiveCfg = Release|Any CPU
 		{9A2197A7-5108-445B-8B54-4B46E525A7D3}.Release|x86.Build.0 = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|x64.Build.0 = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Debug|x86.Build.0 = Debug|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|x64.ActiveCfg = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|x64.Build.0 = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|x86.ActiveCfg = Release|Any CPU
-		{9AF6F544-FB57-432B-8CCF-C32B65B9F82D}.Release|x86.Build.0 = Release|Any CPU
 		{D55E920A-8B56-43AD-A799-6E6154193427}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D55E920A-8B56-43AD-A799-6E6154193427}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D55E920A-8B56-43AD-A799-6E6154193427}.Debug|x64.ActiveCfg = Debug|Any CPU
